+++ components-api/src/main/scala/pl/touk/nussknacker/engine/api/context/ProcessCompilationError.scala
@@ -93,6 +93,14 @@ object ProcessCompilationError {
   ) extends PartSubGraphCompilationError
       with InASingleNode
 
+  final case class InvalidValidationExpression(
+      message: String,
+      nodeId: String,
+      paramName: String,
+      originalExpr: String
+  ) extends PartSubGraphCompilationError
+      with InASingleNode
+
   object ExpressionParserCompilationError {
 
     def apply(message: String, fieldName: Option[String], originalExpr: String)(
@@ -259,14 +267,17 @@ object ProcessCompilationError {
 
   final case class FragmentOutputNotDefined(id: String, nodeIds: Set[String]) extends ProcessCompilationError
 
-  final case class RequireValueFromEmptyFixedList(paramName: String, nodeIds: Set[String])
-      extends ProcessCompilationError
+  final case class RequireValueFromEmptyFixedList(paramName: String, nodeId: String)
+      extends PartSubGraphCompilationError
+      with InASingleNode
 
-  final case class InitialValueNotPresentInPossibleValues(paramName: String, nodeIds: Set[String])
-      extends ProcessCompilationError
+  final case class InitialValueNotPresentInPossibleValues(paramName: String, nodeId: String)
+      extends PartSubGraphCompilationError
+      with InASingleNode
 
-  final case class UnsupportedFixedValuesType(paramName: String, typ: String, nodeIds: Set[String])
-      extends ProcessCompilationError
+  final case class UnsupportedFixedValuesType(paramName: String, typ: String, nodeId: String)
+      extends PartSubGraphCompilationError
+      with InASingleNode
 
   final case class UnknownFragmentOutput(id: String, nodeIds: Set[String]) extends ProcessCompilationError