+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/compile/nodecompilation/FragmentParameterValidator.scala
@@ -5,8 +5,11 @@ import cats.data.ValidatedNel
 import cats.implicits.toTraverseOps
 import pl.touk.nussknacker.engine.api.NodeId
 import pl.touk.nussknacker.engine.api.context.ProcessCompilationError._
-import pl.touk.nussknacker.engine.api.context.{ProcessCompilationError, ValidationContext}
+import pl.touk.nussknacker.engine.api.context.{PartSubGraphCompilationError, ValidationContext}
+import pl.touk.nussknacker.engine.api.definition.ValidationExpressionParameterValidator
+import pl.touk.nussknacker.engine.api.typed.typing.{Typed, TypingResult}
 import pl.touk.nussknacker.engine.compile.ExpressionCompiler
+import pl.touk.nussknacker.engine.expression.NullExpression
 import pl.touk.nussknacker.engine.graph.expression.Expression
 import pl.touk.nussknacker.engine.graph.node.FragmentInputDefinition.{
   FixedExpressionValue,
@@ -19,14 +22,15 @@ class FragmentParameterValidator(
     expressionCompiler: ExpressionCompiler
 ) {
 
-  def validate(
+  def validateAndGetExpressionValidator(
       fragmentParameter: FragmentParameter,
-      fragmentInputId: String,
       validationContext: ValidationContext // localVariables must include this and other FragmentParameters
-  )(implicit nodeId: NodeId): ValidatedNel[ProcessCompilationError, Unit] = {
-    val unsupportedFixedValuesValidationResult = validateFixedValuesSupportedType(fragmentParameter, fragmentInputId)
+  )(
+      implicit nodeId: NodeId
+  ): ValidatedNel[PartSubGraphCompilationError, List[ValidationExpressionParameterValidator]] = {
+    val unsupportedFixedValuesValidationResult = validateFixedValuesSupportedType(fragmentParameter)
 
-    val fixedValuesListValidationResult = validateFixedValuesList(fragmentParameter, fragmentInputId)
+    val fixedValuesListValidationResult = validateFixedValuesList(fragmentParameter)
 
     val fixedExpressionsValidationResult = validateFixedExpressionValues(
       fragmentParameter.initialValue,
@@ -35,27 +39,77 @@ class FragmentParameterValidator(
       fragmentParameter.name
     )
 
+    val validationExpressionResult =
+      compileExpressionValidator(fragmentParameter, validationContext(fragmentParameter.name))
+
     List(
       unsupportedFixedValuesValidationResult,
       fixedValuesListValidationResult,
-      fixedExpressionsValidationResult
-    ).sequence.map(_ => ())
+      fixedExpressionsValidationResult,
+      validationExpressionResult
+    ).sequence.map(_.flatten)
   }
 
-  private def validateFixedValuesSupportedType(fragmentParameter: FragmentParameter, fragmentInputId: String) =
+  private def compileExpressionValidator(
+      fragmentParameter: FragmentParameter,
+      typ: TypingResult
+  )(implicit nodeId: NodeId) =
+    fragmentParameter.valueCompileTimeValidation.map { expr =>
+      expressionCompiler
+        .compile(
+          expr.validationExpression,
+          fieldName = Some(fragmentParameter.name),
+          validationCtx = ValidationContext(
+            Map(ValidationExpressionParameterValidator.variableName -> typ)
+          ), // TODO in the future, we'd like to support more references, see ValidationExpressionParameterValidator
+          expectedType = Typed[Boolean],
+        )
+        .leftMap(_.map {
+          case e: ExpressionParserCompilationError =>
+            InvalidValidationExpression(
+              e.message,
+              nodeId.id,
+              fragmentParameter.name,
+              e.originalExpr
+            )
+          case e => e
+        })
+        .andThen {
+          _.expression match {
+            case _: NullExpression =>
+              invalidNel(
+                InvalidValidationExpression(
+                  "Validation expression cannot be blank",
+                  nodeId.id,
+                  fragmentParameter.name,
+                  expr.validationExpression.expression
+                )
+              )
+            case expression =>
+              Valid(
+                ValidationExpressionParameterValidator(
+                  expression,
+                  fragmentParameter.valueCompileTimeValidation.flatMap(_.validationFailedMessage)
+                )
+              )
+          }
+        }
+    }.sequence
+
+  private def validateFixedValuesSupportedType(fragmentParameter: FragmentParameter)(implicit nodeId: NodeId) =
     fragmentParameter.valueEditor match {
       case Some(_) =>
         if (List(FragmentClazzRef[java.lang.Boolean], FragmentClazzRef[String]).contains(fragmentParameter.typ)) {
-          Valid(())
+          Valid(None)
         } else
           invalidNel(
             UnsupportedFixedValuesType(
               fragmentParameter.name,
               fragmentParameter.typ.refClazzName,
-              Set(fragmentInputId)
+              nodeId.id
             )
           )
-      case None => Valid(())
+      case None => Valid(None)
     }
 
   private def validateFixedExpressionValues(
@@ -101,24 +155,23 @@ class FragmentParameterValidator(
         fixedValuesList,
         Some(FixedValuesListFieldName)
       )
-    ).sequence.map(_ => ())
+    ).sequence.map(_ => None)
   }
 
   private def validateFixedValuesList(
       fragmentParameter: FragmentParameter,
-      fragmentInputId: String
-  ): ValidatedNel[ProcessCompilationError, Unit] =
+  )(implicit nodeId: NodeId) =
     fragmentParameter.valueEditor match {
       case Some(valueEditor) if !valueEditor.allowOtherValue =>
         List(
           if (valueEditor.fixedValuesList.isEmpty)
-            invalidNel(RequireValueFromEmptyFixedList(fragmentParameter.name, Set(fragmentInputId)))
+            invalidNel(RequireValueFromEmptyFixedList(fragmentParameter.name, nodeId.id))
           else Valid(()),
           if (initialValueNotPresentInPossibleValues(fragmentParameter))
-            invalidNel(InitialValueNotPresentInPossibleValues(fragmentParameter.name, Set(fragmentInputId)))
+            invalidNel(InitialValueNotPresentInPossibleValues(fragmentParameter.name, nodeId.id))
           else Valid(())
-        ).sequence.map(_ => ())
-      case _ => Valid(())
+        ).sequence.map(_ => None)
+      case _ => Valid(None)
     }
 
   private def initialValueNotPresentInPossibleValues(