+++ docs/Changelog.md
@@ -29,7 +29,11 @@
 * [#4887](https://github.com/TouK/nussknacker/pull/4887) New parameter validator - `ValidationExpressionParameterValidator` that allows to use SpEL (or any other) expression to validate parameters
 * [#5077](https://github.com/TouK/nussknacker/pull/5077) Add an option to set schema on connections in SQL enricher
 * [#5059](https://github.com/TouK/nussknacker/pull/5059) [#5100](https://github.com/TouK/nussknacker/pull/5100) [#5103](https://github.com/TouK/nussknacker/pull/5103) Breaking change: Scenario type to Category mapping become 1-to-1 instead of 1-to-many.
-* [#4978](https://github.com/TouK/nussknacker/pull/4978) Expand FragmentParameter with new fields: hintText, initialValue, required, inputMode
+* [#4978](https://github.com/TouK/nussknacker/pull/4978) [#5241](https://github.com/TouK/nussknacker/pull/5241) Expand `FragmentParameter` with new fields:
+  * `hintText` - shown next to the parameter when using the fragment
+  * `initialValue` - initial value of the parameter (before user changes it)
+  * `required` - whether the parameter is mandatory
+  * `valueCompileTimeValidation` - allows configuration of `ValidationExpressionParameterValidator` for this parameter
 * [#4953](https://github.com/TouK/nussknacker/pull/4953) Improved node validation
 * [#5141](https://github.com/TouK/nussknacker/pull/5141) Security improvement: API endpoints check if user has access rights to Category associated with Processing Type provided in API
 * [#5182](https://github.com/TouK/nussknacker/pull/5182) [#5203](https://github.com/TouK/nussknacker/pull/5203) [#5250](https://github.com/TouK/nussknacker/pull/5250) Component, User and Notification API OpenAPI-based documentation (e.g. `https://demo.nussknacker.io/api/docs`)
@@ -39,7 +43,7 @@
 * [#5233](https://github.com/TouK/nussknacker/pull/5233) Fix: Not expected type: Null runtime error for non-nullable, optional json fields that were not provided by user in sink
 * [#5233](https://github.com/TouK/nussknacker/pull/5233) Added support for schema evolution in kafka using json schema and response sink
 
-* 1.12.5 (1 Dec 2023)
+1.12.5 (1 Dec 2023)
 ------------------------
 * [#5110](https://github.com/TouK/nussknacker/pull/5110) Fix: The compare option doesn't display fragment input properties between the two versions